services/machine_interface/requirements.txt
kafka-python
pymodbus
